separate apis
add progress bar
