+++ ozerpan_ercom_sync/public/js/sales_order/sales_order_list.js
@@ -58,14 +58,12 @@ function callUploadBomApi(values) {
   loading.show();
 
   frappe.call({
-    method: "ozerpan_ercom_sync.api.update_bom",
+    method: "ozerpan_ercom_sync.custom_api.sales_order.update_bom",
     args: {
       file_url: values.file,
     },
     freeze: true,
     callback: (r) => {
-      loading.hide();
-
       if (r.message) {
         console.log(r.message);
         frappe.msgprint({
@@ -76,7 +74,6 @@ function callUploadBomApi(values) {
       }
     },
     error: (r) => {
-      loading.hide();
       frappe.msgprint({
         title: __("Error"),
         indicator: "red",
@@ -87,31 +84,10 @@ function callUploadBomApi(values) {
 }
 
 function callSyncErcomApi() {
-  let loading = new frappe.ui.Dialog({
-    title: __("Processing"),
-    primary_action_label: __("Hide"),
-    primary_action: () => loading.hide(),
-  });
-
-  loading.$body.html(`
-    <div class="progress">
-      <div class="progress-bar progress-bar-striped active" role="progressbar" style="width: 100%">
-        <span>${__("Processing Customer Sync... Please wait.")}</span>
-      </div>
-    </div>
-    <div class="margin-top">
-      <p class="text-muted">${__("This may take a few minutes. Please do not close this window.")}</p>
-    </div>
-    `);
-
-  loading.show();
-
   frappe.call({
-    method: "ozerpan_ercom_sync.api.sync_ercom",
+    method: "ozerpan_ercom_sync.custom_api.sync_ercom.sync_ercom",
     freeze: true,
     callback: (r) => {
-      loading.hide();
-
       if (r.message) {
         console.log(r.message);
         frappe.msgprint({
@@ -122,7 +98,6 @@ function callSyncErcomApi() {
       }
     },
     error: (r) => {
-      loading.hide();
       frappe.msgprint({
         title: __("Error"),
         indicator: "red",
